Add set-area and set-watcher for container level
* add attribute watcher-name for tag container
* add in link-type enum set-area, refresh-area and set-watcher,
refresh-watcher
* add in xxx-event-update-area, in event-type enum set-area,
refresh-area and set-watcher, refresh-watcher

Concept is:
1. container is the main piece of dynamics screen (update a part of
screen on event), a new attribute is added watcher-name, and existing
one auto-update-target is using as target

2. There are two/four new "action" usable by link - hyperlink
 * set-area (and refresh-area)
 * set-watcher (and refresh-watcher)

3. set-area is used to update a containerId, and, most of the time, uses
only for included containerId
set-watcher put some "parameters" in a watcher-name

4. when the content of a whatcher-name change, all containerId which
watch it, use their auto-update-target as target and watcher-name
content as parameters to update the container content.

5. refresh-area is similar with set-area but target used will be the
auto-update-target define at the container level and, if a watcher-name
is define for this container, it's content will be used as parameters
(for solving {} in target or as parameters if no {} in target.